+++ src/jsMain/kotlin/com/foundryvtt/core/documents/TokenDocument.kt
@@ -1,6 +1,14 @@
 package com.foundryvtt.core.documents
 
-import com.foundryvtt.core.*
+import com.foundryvtt.core.Actor
+import com.foundryvtt.core.AnyObject
+import com.foundryvtt.core.CreateDocumentCallback
+import com.foundryvtt.core.DeleteDocumentCallback
+import com.foundryvtt.core.HooksEventListener
+import com.foundryvtt.core.PreCreateDocumentCallback
+import com.foundryvtt.core.PreDeleteDocumentCallback
+import com.foundryvtt.core.PreUpdateDocumentCallback
+import com.foundryvtt.core.UpdateDocumentCallback
 import com.foundryvtt.core.abstract.DataModel
 import com.foundryvtt.core.abstract.DatabaseDeleteOperation
 import com.foundryvtt.core.abstract.DatabaseUpdateOperation
@@ -157,8 +165,8 @@ external class TokenDocument : Document {
     var height: Int
     var texture: TextureData
     var hexagonalShape: Int
-    var x: Int
-    var y: Int
+    var x: Double
+    var y: Double
     var elevation: Int
     var sort: Int
     var locked: Boolean
